+++ R/zotero_match.R
@@ -16,7 +16,7 @@
 #'   NULL
 #' @param polite Will use an email stored in `.Renviron`, Default: TRUE
 #' @param silent Running silent, running deep, Default: FALSE
-#' @param prefer.semantic Prefer metadata from Semantic Scholar, Default: FALSE
+#' @param use.semantic Prefer metadata from Semantic Scholar, Default: FALSE
 #' @param log A list for storing log elements, Default: list()
 #' @return A Zotero-type matrix (tibble) if match is found otherwise NULL
 #' @details Please see
@@ -60,7 +60,7 @@ ZoteroMatch <- \(title,
                  external.data = NULL,
                  polite = TRUE,
                  silent = FALSE,
-                 prefer.semantic = FALSE,
+                 use.semantic = FALSE,
                  log = list()) {
 
   # Visual bindings
@@ -368,7 +368,7 @@ ZoteroMatch <- \(title,
 
     # Find metadata from Crossref using DOI
     doi <- haystack.data[best.match,"DOI"]
-    doi <- ZoteroDoi(doi, prefer.semantic = prefer.semantic)
+    doi <- ZoteroDoi(doi, use.semantic = use.semantic)
     result <- doi$data
     log <- append(log, doi$log)
     if (!is.null(result)) {